Add ioq:bypass/2 API
Given a message and a priority, return `true` if that operation would be
bypassed. This is used by the parallel preads commit [1] with the cfile handle
when we know the request would be bypassed by the IOQ.

If/when we figure out how to teach the IOQ server to call a regular MFA, as
opposed to always sending a '$gen_call' message, we won't have a need for
this and can remove this API.

While at it, do some minor test housekeeping and use the standard `?TDEF_FE`
helper macro to remove the eunit `?_test/begin` boilerplate code.

[1] https://github.com/apache/couchdb/pull/5399