Make sure to handle `other` priority in ioq's new bypass/2 API
It's a new API to check if the request would be bypassed by the IOQ. IOQ2
server didn't handle that clause properly and it would throw function_clause
errors as a result.

IOQ1 apparently handles any value for the priority so we ensure to do the same.
Normally it would be `undefined` though but we'll go with IOQ1's approach and
accept any junk there since it's the "other" category.